Risolti problemi compilazione BeaconManagerTest
