+++ app/build.gradle
@@ -79,8 +79,4 @@ dependencies {
     compile 'com.google.code.findbugs:jsr305:2.0.1'
 
 
-    testCompile 'junit:junit:4.12'
-    compile 'junit:junit:4.12'
-    compile 'junit:junit:4.12'
-
 }
