+++ app/src/androidTest/java/com/leaf/clips/model/beacon/BeaconManagerAndroidTest.java
@@ -1,9 +1,10 @@
 package com.leaf.clips.model.beacon;
 /**
  * @author Federico Tavella
- * @version 0.01
+ * @version 0.03
  * @since 0.00
  */
+
 
 import android.content.Context;
 import android.content.Intent;
@@ -14,6 +15,7 @@ import android.support.test.rule.ServiceTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.InstrumentationTestCase;
 import android.test.suitebuilder.annotation.SmallTest;
+
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
@@ -23,11 +25,11 @@ import org.junit.runner.RunWith;
 import java.util.concurrent.TimeoutException;
 
 /**
- * Unit test 40
+ * Unit test 39 & 40
  */
 @RunWith(AndroidJUnit4.class)
 @SmallTest
-public class BeaconManagerAndroidTest40 extends InstrumentationTestCase{
+public class BeaconManagerAndroidTest extends InstrumentationTestCase{
 
     private BeaconRanger ranger;
 
@@ -35,7 +37,6 @@ public class BeaconManagerAndroidTest40 extends InstrumentationTestCase{
 
     private Context context = InstrumentationRegistry.getTargetContext();
 
-    @Rule
     public final ServiceTestRule mServiceRule = new ServiceTestRule();
 
     @Before
@@ -45,6 +46,18 @@ public class BeaconManagerAndroidTest40 extends InstrumentationTestCase{
         IBinder binder = mServiceRule.bindService(intent);
         ranger = ((BeaconManagerAdapter.LocalBinder) binder).getService();
 
+    }
+
+    @Test
+    public void shouldSwitchToBackgroundMode(){
+
+        ranger.setBackgroundMode(true);
+
+        Assert.assertEquals(true, ranger.isBackground());
+
+        ranger.setBackgroundMode(false);
+
+        Assert.assertEquals(false, ranger.isBackground());
     }
 
     @Test
